chapter2: adjust fmp format wording
We require in-band firmware updates in the "Firmware Management Protocol
Data Capsule Structure" format. This is a bit misleading as this format
described in UEFI § 23.3.2 is only about the capsule body.

Refer to the Firmware Management Protocol format instead to avoid the
confusion, remove the unnecessary quotation marks and introduce the FMP
acronym.

Signed-off-by: Vincent Stehlé <vincent.stehle@arm.com>